==> notebooks/1-EDA.ipynb <==
# Imports

import pandas as pd

# Exploratory Data Analysis

Let us look at the data.

## Users Data

users_columns = ['user_id', 'age', 'sex', 'occupation', 'zip_code']

df_users = pd.read_csv(
    '../data/raw/ml-100k/u.user',
    sep='|',
    encoding='latin-1',
    index_col=0,
    names=users_columns
)

print(df_users.shape)
df_users.head()

## Ratings Data

ratings_columns = ['user_id', 'movie_id', 'rating', 'unix_timestamp']

df_ratings = pd.read_csv(
    '../data/raw/ml-100k/u.data',
    sep='\t',
    encoding='latin-1',
    index_col=0,
    names=ratings_columns
)

print(df_ratings.shape)
df_ratings.head()

## Items Data

items_columns = ['movie_id', 'title', 'release_date', 'video_release_date', 'IMDb URL',
                 'unknown', 'Action', 'Adventure', 'Animation', 'Children\'s', 'Comedy', 'Crime',
                 'Documentary', 'Drama', 'Fantasy', 'Film-Noir', 'Horror', 'Musical', 'Mystery',
                 'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western']

df_items = pd.read_csv(
    '../data/raw/ml-100k/u.item',
    sep='|',
    encoding="latin-1",
    index_col=0,
    names=items_columns
)

print(df_items.shape)
df_items.head()

## Remove Poor and Unused Data

df_items[df_items['unknown']==1]

df_ratings[df_ratings['movie_id'].isin([267, 1373])]

There are two movies with unknown genre. We see there are few ratings for them, so let us remove these movies and ratings from the dataset.

We will not use some columns, so let us also leave only the ones needed to build the user vectors.

# remove unused columns
df_items = df_items.drop(columns=['release_date', 'video_release_date', 'IMDb URL'])
df_ratings = df_ratings.drop(columns=['unix_timestamp'])
df_users = df_users.drop(columns=['zip_code'])

# remove unknown genre films
print(f'before removal: {df_items.shape[0]} movies and {df_ratings.shape[0]} ratings')
unknown_genre_movies_ids = df_items[df_items['unknown'] == 1].index
df_items = df_items.drop(unknown_genre_movies_ids)
df_ratings = df_ratings.loc[~df_ratings['movie_id'].isin(unknown_genre_movies_ids)]
print(f'after removal: {df_items.shape[0]} movies and {df_ratings.shape[0]} ratings')

# re-index the user_id column
df_ratings.index = df_ratings.index - 1
df_users.index = df_users.index - 1

Let us save the preprocessed data:

df_items.to_csv('../data/interim/preprocessed/items.csv')
df_ratings.to_csv('../data/interim/preprocessed/ratings.csv')
df_users.to_csv('../data/interim/preprocessed/users.csv')
