==> notebooks/2-users-vectors.ipynb <==
# Imports

import numpy as np
import pandas as pd

from sklearn.metrics.pairwise import cosine_similarity

# User Vectors

Let us build the user vectors using the following info:

- age (int)
- sex (boolean)
- occupation (one-hot encoded)
- genre favor vector (sum of `user's rating to a movie * one-hot vector of the movie genres` for all the user ratings)

### Read the tables

df_users = pd.read_csv(
    '../data/interim/preprocessed/users.csv',
    index_col=0
)

print(df_users.shape)
df_users.head()

df_items = pd.read_csv(
    '../data/interim/preprocessed/items.csv',
    index_col=0
)

print(df_items.shape)
df_items.head()

genres = ['Action', 'Adventure', 'Animation', 'Children\'s', 'Comedy', 'Crime',
          'Documentary', 'Drama', 'Fantasy', 'Film-Noir', 'Horror', 'Musical', 'Mystery',
          'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western']

df_ratings = pd.read_csv(
    '../data/interim/preprocessed/ratings.csv',
    index_col=0
)

df_ratings = df_ratings.merge(df_items[genres], how='inner', left_on='movie_id', right_index=True)

print(df_ratings.shape)
df_ratings.head()

### Add genre favor data

for genre in genres:
    df_ratings[genre] = df_ratings['rating'] * df_ratings[genre]

user_genre_sum = df_ratings.groupby('user_id')[genres].sum()#.reset_index()

# normalize each user's genre fevor vector independently
user_genre_sum[genres] = user_genre_sum[genres].div(user_genre_sum[genres].max(axis=1), axis=0)

# merge the genre favors into df_users
df_users = pd.merge(df_users, user_genre_sum, how='left', on='user_id')

# fill nans with 0 (users without ratings for a genre)
df_users[genres] = df_users[genres].fillna(0)

df_users.head()

# Encode age, sex, and occupation

print(df_users['age'].max())

# take as a maximum age
max_age = 100
df_users['age'] /= max_age

# one-hot encode the sex column
df_users = pd.get_dummies(df_users, columns=['sex'], prefix='sex').drop(columns=['sex_F'])

# one-hot encode the occupation column
df_users = pd.get_dummies(df_users, columns=['occupation'], prefix='occ')

df_users

# Cosine similarity

Let us compute the cosine similarities between user vectors. Then for each user save the list of other user ids sorted by descending of their similarity.

# extract user vectors as numpy arrays
user_vectors = df_users[df_users.columns].to_numpy()

# compute cosine similarity
similarity_matrix = cosine_similarity(user_vectors, user_vectors)

# dict to store the similarity info
users_similarity = {'user_id': [], 'similar_ids': []}

for i in range(len(df_users)):
    user_ids = range(len(similarity_matrix[i]))

    # get the list of user indices sorted by the cosine similarity with user i
    similar_indices = sorted(user_ids, key=lambda j: similarity_matrix[i][j], reverse=True)
    similar_indices = [(j, similarity_matrix[i][j]) for j in similar_indices if j != i]

    # save the similarity list
    users_similarity['user_id'].append(i)
    users_similarity['similar_ids'].append(similar_indices)

df_users_similarity = pd.DataFrame(users_similarity).set_index('user_id')
df_users_similarity

Max and min cosine similarity values:

print(similarity_matrix.max(), similarity_matrix.min())

Let us choose the thresholds for the cosine similarity:

threshold_similarity = 0.9
threshold_count = 2

np.sum(np.count_nonzero(similarity_matrix > threshold_similarity, axis=1) < threshold_count)

As you can see, there are only 17 out of 943 users who have less than 2 users with similarity score above 0,9. We can use this 0,9 threshold to filter the similar users that will influence the rating recommended by SVD algorithm.

We can now save the user vectors and similarity table:

df_users.to_csv('../data/interim/users/user_vectors.csv')
df_users_similarity.to_csv('../data/interim/users/users_similarity.csv')
