==> notebooks/3-recsys.ipynb <==
# Imports and Data Import

import numpy as np
import pandas as pd
import json
import sys

sys.path.insert(0, "../")

from models.recsys import recommend_svd, recommend_hybrid, print_recs
from surprise import Dataset, Reader, SVD, dump
from surprise.model_selection import cross_validate

df_ratings = pd.read_csv(
    '../data/interim/preprocessed/ratings.csv',
    index_col=0
).reset_index()

print(df_ratings.shape)
df_ratings.head()

df_items = pd.read_csv(
    '../data/interim/preprocessed/items.csv',
    index_col=0
)

print(df_items.shape)
df_items.head()

# SVD

reader = Reader(rating_scale=(1, 5))

columns = ['user_id', 'movie_id', 'rating']
ratings_data = Dataset.load_from_df(df_ratings[columns], reader)

svd = SVD()

cross_validate(svd, ratings_data, measures=['RMSE', 'MAE'], cv=5, verbose=True)

# RecSys

A recsys implementation with two options:

- SVD only
- hybrid: SVD + cosine similarity-based recommendations (considering the users data as well)

df_users_similarity = pd.read_csv(
    '../data/interim/users/users_similarity.csv',
    index_col=0
)

print(df_users_similarity.shape)
df_users_similarity.head()

movies_ids = df_items.index.tolist()

test_user_id = 2

Recommendations with pure SVD algorithm:

recs_best, recs_all = recommend_svd(
    user_id=test_user_id,
    df_ratings=df_ratings,
    movies_ids=movies_ids,
    svd=svd,
    num_of_recs=5
)
print_recs(df_items, recs_best)

Recommendations with hybrid system (SVD + users similarity):

recs_best, recs_all = recommend_hybrid(
    recs=recs_all,
    user_id=test_user_id,
    df_ratings=df_ratings,
    df_users_similarity=df_users_similarity,
    similarity_threshold=0.9,
    num_of_recs=5
)
print_recs(df_items, recs_best)

Serialize the SVD model:

dump.dump('../models/svd_dump_file', algo=svd)
